Allow cameras to go upside-down. Fixed viewport warnings. Unknown cause of P2 starting upside-down on sphere.

+++ HardLight/src/Rendering/Viewports.cpp
@@ -2,7 +2,7 @@
 
 // Math is pretty difficult to dynamically calculate this on the fly so we will load based on cases
 // Will not initialize camera object. That will be needed to be created later on
-vector<Viewports::Viewport> Viewports::generate_viewports(int numViews, float width, float height)
+vector<Viewports::Viewport> Viewports::generate_viewports(int numViews, int width, int height)
 {
 	vector<Viewports::Viewport> viewports;
 
@@ -27,16 +27,16 @@ vector<Viewports::Viewport> Viewports::generate_viewports(int numViews, float wi
 
 			// Top frame
 			viewport1.x = 0;
-			viewport1.y = int(height/2);
+			viewport1.y = height/2;
 			viewport1.width = width;
-			viewport1.height = height/2.0f;
+			viewport1.height = height/2;
 			viewport1.id = PLAYER1;
 
 			// Bottom frame
 			viewport2.x = 0;
 			viewport2.y = 0;
 			viewport2.width = width;
-			viewport2.height = height/2.0f;
+			viewport2.height = height/2;
 			viewport2.id = PLAYER2;
 
 			viewports.push_back(viewport1);
@@ -51,23 +51,23 @@ vector<Viewports::Viewport> Viewports::generate_viewports(int numViews, float wi
 
 			// Top left frame
 			viewport1.x = 0;
-			viewport1.y = int(height/2);
-			viewport1.width = width/2.0f;
-			viewport1.height = height/2.0f;
+			viewport1.y = height/2;
+			viewport1.width = width/2;
+			viewport1.height = height/2;
 			viewport1.id = PLAYER1;
 
 			// Top right frame
-			viewport2.x = int(width/2);
-			viewport2.y = int(height/2);
-			viewport2.width = width/2.0f;
-			viewport2.height = height/2.0f;
+			viewport2.x = width/2;
+			viewport2.y = height/2;
+			viewport2.width = width/2;
+			viewport2.height = height/2;
 			viewport2.id = PLAYER2;
 
 			// Bottom frame
 			viewport3.x = 0;
 			viewport3.y = 0;
 			viewport3.width = width;
-			viewport3.height = height/2.0f;
+			viewport3.height = height/2;
 			viewport3.id = PLAYER3;
 
 			viewports.push_back(viewport1);
@@ -84,30 +84,30 @@ vector<Viewports::Viewport> Viewports::generate_viewports(int numViews, float wi
 
 			// Top left frame
 			viewport1.x = 0;
-			viewport1.y = int(height/2);
-			viewport1.width = width/2.0f;
-			viewport1.height = height/2.0f;
+			viewport1.y = height/2;
+			viewport1.width = width/2;
+			viewport1.height = height/2;
 			viewport1.id = PLAYER1;
 
 			// Top right frame
-			viewport2.x = int(width/2);
-			viewport2.y = int(height/2);
-			viewport2.width = width/2.0f;
-			viewport2.height = height/2.0f;
+			viewport2.x = width/2;
+			viewport2.y = height/2;
+			viewport2.width = width/2;
+			viewport2.height = height/2;
 			viewport2.id = PLAYER2;
 
 			// Bottom left frame
 			viewport3.x = 0;
 			viewport3.y = 0;
-			viewport3.width = width/2.0f;
-			viewport3.height = height/2.0f;
+			viewport3.width = width/2;
+			viewport3.height = height/2;
 			viewport3.id = PLAYER3;
 			
 			// Bottom right frame
-			viewport4.x = int(width/2);
+			viewport4.x = width/2;
 			viewport4.y = 0;
-			viewport4.width = width/2.0f;
-			viewport4.height = height/2.0f;
+			viewport4.width = width/2;
+			viewport4.height = height/2;
 			viewport4.id = PLAYER4;
 
 			viewports.push_back(viewport1);
